third_party/WebKit/LayoutTests/http/tests/security/sandboxed-iframe-blocks-access-from-parent.html

<!doctype html>
<html>
<head>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFramesAsText();
        }

        function test() {
            var test = document.getElementById('testframe');
            try {
                var location = test.contentWindow.location.href;
            } catch (e) {
                alert("PASS: Cross-origin access to the Location object threw an exception.");
            }
        }
    </script>
</head>

<body>
    <p>Sandboxing a frame puts it into a unique origin by default, which the
    containing document shouldn't have script access to. This test passes if an
    exception is thrown upon the access violation.</p>

    <iframe src="resources/blank.html" id="testframe" sandbox onload="test();"></iframe>
</body>
</html>
