third_party/WebKit/LayoutTests/http/tests/security/sandboxed-iframe-blocks-access-from-parent-expected.txt

CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "http://127.0.0.1:8000" from accessing a frame at "http://127.0.0.1:8000".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
ALERT: PASS: Cross-origin access to the Location object threw an exception.
Sandboxing a frame puts it into a unique origin by default, which the containing document shouldn't have script access to. This test passes if an exception is thrown upon the access violation.



--------
Frame: 'testframe'
--------
